Restore wmf/1.35.0-wmf.5 to shippable state.
There were some mistakes on the reading web team's part as part of the
wmf/1.35.0-wmf.5 deploy. This restores the state to what we wanted to
go out on the train.

This is a combination of 3 commits:
* This reverts commit 3a6729841b70d9b04c0fc54c82b4abfd555cbc70. ( "Remove eventbus usage in talk.js")
* This reverts commit 2d579183c9d36e46bdd92361882d0cf7bc865cee. ("Render talk page as a tab instead of an overlay")
* This adds commit b4f03d475fd484b7f674cb9c0d4e0554d16a1649
("Do not jump to the top of the page when the reference drawer is closed")

Depends-On: Ieb1ffd7a64a79a6af2218a93e6d7e2d903baa0e2
Change-Id: Ia1d5adf9759f5da5f7c2b6bc09bfa9fb7a9a4d32