Remove double setIcon() call for Contributions element on Toolbar
Fix Contributions element classes by not using ::create named
constructor. The constructor was provided for easier object creation
when both icon and click tracking is the same as element name.
The Contributions icon is not that case, as both tracking code and
icon are definited differently than the name, therefore we need
to initailize class, and set required attributes manually.

Change-Id: Ia91bbdcfd737a443a05fd4f61853fe8bd200e56c