Remove double setIcon() call for the History element on Toolbar
Fix History element classes by not using ::create named
constructor. The constructor was provided for easier object creation
when both icon and click tracking is the same as element name.
The History icon is not that case, as both tracking code and
icon are definited differently than the name, therefore we need
to initailize class, and set required attributes manually.

Change-Id: I7b74e8a7e976d6c1e47f15438f75a7de2f73f154