Update Minerva's contributors
I want to reflect the contributions of the
existing members of the team that
maintain the skin.

All the added names are listed as the main contributors
on the github with consistent contributions
of at least 2 years
https://github.com/wikimedia/mediawiki-skins-MinervaNeue/graphs/contributors

Alex has overseen much of the effort in our mobile skin over the recent years
Although we don't traditionally add designers to the authors section I think
it makes sense that he should be listed.

Change-Id: Ia2a37a6988a2712151ae0c21f9e31a11310a8a53