Don't add the "minerva--history-page-action-enabled" class to <body> for non-Minerva skins like MonoBook etc.
Hooks are global, hence any and all skin-specific OutputPageBodyAttributes hooks need to check first that the correct skin (or skins) is used, otherwise the class or classes get applied for *all* skins when only one skin is supposed to be targeted.

Also fixed a tiny code style issue in the same hook subscriber while at it.

Change-Id: I626b0e050c614687b31f1ffbaea92e371dee4574