ServiceWiring: Add fallback for when title is null
Services shouldn't rely on global state. The service should have a
method for setting the title manually, and callers to the service should
be updated to pass in the title (or other context items) as needed. As
is, this code causes the test for MediaWikiServicesTest.php to fail; it
appears to only pass in CI accidentally due to pollution of global state
from other tests that run before it.

Bug: T297068
Change-Id: I39c780d36583deb8b2edd7b12b425e949bd09cd1