Remove redundant styling rules and fix center/none halign for figures
The rules in mediawiki.skinning.content.parsoid were made redundant by
25890e3ece5921def5f936de8c4ae8bbe85c4b5d.  There are some more rules
there that should be audited to see if they're still necessary or can be
moved to a more general location now that wgParserEnableLegacyMediaDOM
is being disabled and they apply more broadly than just when Parsoid
styles are requested.

Missing rules for mw-halign-center and mw-halign-none are added.

Bug: T318300
Change-Id: I34fb1ce76de24eec3b136dc6dc9ab4e07e36b809