selenium: Enable one skipped test, delete one failing test
'Broken redirects' was disabled in 69a100cb45f47dc97331f4e8613488ff1aa86db5
'should be creatable' was disabled in 5e0dd88af04e20a3356e93939aa46fc2eadd1e37

'should be creatable' is failing so it is deleted.
'Broken redirects' is passing so it is enabled.

To be sure that the passing test is not flaky, the CI has ran it 100
times, it never failed.

Bug: T280652
Change-Id: I34a863dcbd126c2da8a98624e6ba1fa916eac0cf