Blank skin styles for login and create account
In 2049c81e776f these were marked as being targeted
on mobile, meaning they would start to load on mobile site.

Minerva doesn't blank skin styles by default so gets the "default"
set of styles.

Since these have a negative effect this is not likely desired
so we should blank the styles to restore how it used to be for
now.

Bug: T327661
Change-Id: Ifc6cdb4a1d756edf23396b462ba4447f676ab737