Revert "Add custom styles for temporary user links"
This reverts commit 52273510fe3854ebf34a947315ae45c1b41f87c9.

Reason for revert: Web team is trying to limit the amount of differences between Minerva and Vector codebases, using shared styles
in core where-ever possible. Since this rule is already defined in core, we should load it from there. This requires a slight change to
the loading mechanism, which I've marked as a dependency.

Change-Id: Id31a6b24bdd4f745e194d77cf7b97faf6ae7e6de
Depends-On: I11c54b58873523f7b72cca67b82efdd681562508